+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add support for `visibility` in `Groups::all`
 * Add support for personal access tokens
 * Add support for `job_inputs` and `job_variables_attributes` in `Jobs::play`
+* Add support for filters in `Projects::projectAccessTokens`
 
 
 ## [12.0.0] - 2025-02-23